+++ config/cdnify.js
@@ -1,0 +1,8 @@
+'use strict';
+
+// Replace Google CDN references
+module.exports = {
+  dist: {
+    html: ['<%= yeoman.dist %>/client/*.html']
+  }
+};
