+++ config/istanbul_check_coverage.js
@@ -1,0 +1,15 @@
+'use strict';
+
+module.exports = {
+  default: {
+    options: {
+      coverageFolder: 'coverage/**',
+      check: {
+        lines: 80,
+        statements: 80,
+        branches: 80,
+        functions: 80
+      }
+    }
+  }
+};
