+++ config/karma.js
@@ -1,0 +1,9 @@
+'use strict';
+
+// Test settings
+module.exports = {
+  unit: {
+    configFile: 'karma.conf.js',
+    singleRun: true
+  }
+};
