+++ config/nodeInspector.js
@@ -1,0 +1,10 @@
+'use strict';
+
+// Debugging with node inspector
+module.exports = {
+  custom: {
+    options: {
+      'web-host': 'localhost'
+    }
+  }
+};
