+++ config/open.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  server: {
+    url: 'http://localhost:<%= express.options.port %>'
+  }
+};
