+++ config/protractor.js
@@ -1,0 +1,14 @@
+'use strict';
+
+module.exports = {
+  options: {
+    configFile: 'protractor.conf.js'
+  },
+  chrome: {
+    options: {
+      args: {
+        browser: 'chrome'
+      }
+    }
+  }
+};
