+++ config/sass.js
@@ -1,0 +1,13 @@
+'use strict';
+
+// Compiles Sass to CSS
+module.exports = {
+  server: {
+    options: {
+      compass: false
+    },
+    files: {
+      '.tmp/app/app.css' : '<%= yeoman.client %>/app/app.scss'
+    }
+  }
+};
