+++ package.json
@@ -90,6 +90,8 @@
     "karma-requirejs": "~0.2.2",
     "karma-script-launcher": "~0.1.0",
     "karma-spec-reporter": "~0.0.20",
+    "load-grunt-configs": "^0.4.3",
+    "load-grunt-tasks": "^3.4.1",
     "mocha": "^2.2.5",
     "open": "~0.0.4",
     "phantomjs-prebuilt": "^2.1.4",
